requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-ai-formrecognizer
azure-core
psycopg2-binary